uint64_t

+++ code/DeepSleep/ExternalWakeUp_ext1_with_GPIO/ExternalWakeUp_ext1_with_GPIO.ino
@@ -49,7 +49,7 @@ void print_wakeup_reason(){
 Method to print the GPIO that triggered the wakeup
 */
 void print_GPIO_wake_up(){
-  int GPIO_reason = esp_sleep_get_ext1_wakeup_status();
+  uint64_t GPIO_reason = esp_sleep_get_ext1_wakeup_status();
   Serial.print("GPIO that triggered the wake up: GPIO ");
   Serial.println((log(GPIO_reason))/log(2), 0);
 }
